+++ src/voidgl/base-layers/arcs/arc-layer.fs
@@ -1,0 +1,11 @@
+precision highp float;
+
+${picking}
+varying vec4 vertexColor;
+${extendHeader}
+
+void main() {
+  setColor(vertexColor);
+  gl_FragColor = vertexColor;
+  ${extend}
+}